updated podspec

+++ TTFortuneWheelSample/ViewController.swift
@@ -25,7 +25,7 @@ class ViewController: UIViewController {
         spinningWheel.slices = slices
         spinningWheel.equalSlices = true
         spinningWheel.frameStroke.width = 0
-        spinningWheel.titleRotation = Float.pi
+        spinningWheel.titleRotation = CGFloat.pi
         spinningWheel.slices.enumerated().forEach { (pair) in
             let slice = pair.element as! CarnivalWheelSlice
             let offset = pair.offset
